old01/requirements.txt
# o3-pro チャットボットツールキット 依存関係

# Azure OpenAI接続
openai>=1.5.0

# Azure認証  
azure-identity>=1.15.0

# 設定管理
python-dotenv>=1.0.0